# server_dotnet/tests/Contracts/openapi-v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "YF supplier collaboration API",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/admin/system/configs": {
      "get": {
        "tags": [
          "Yf.Api.Tests"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SystemConfigResponse"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Yf.Api.Tests"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigBatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/system/mail-status": {
      "get": {
        "tags": [
          "Yf.Api.Tests"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailStatusResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/system/mail-settings": {
      "get": {
        "tags": [
          "Yf.Api.Tests"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SmtpSettingsView"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Yf.Api.Tests"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SmtpSettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SmtpSettingsView"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/audit-logs": {
      "get": {
        "tags": [
          "Yf.Api.Tests"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseOfAuditLogResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiErrorResponse": {
        "required": [
          "code",
          "message"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AuditLogResponse": {
        "required": [
          "action",
          "actorName",
          "actorNameSource",
          "createdAt",
          "detail",
          "employeeNo",
          "id",
          "ip",
          "targetId",
          "targetName",
          "targetNameSource",
          "targetType",
          "userId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "employeeNo": {
            "type": "string",
            "nullable": true
          },
          "action": {
            "type": "string"
          },
          "targetType": {
            "type": "string",
            "nullable": true
          },
          "targetId": {
            "type": "string",
            "nullable": true
          },
          "detail": {
            "nullable": true
          },
          "ip": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "actorName": {
            "type": "string",
            "nullable": true
          },
          "targetName": {
            "type": "string",
            "nullable": true
          },
          "actorNameSource": {
            "type": "string"
          },
          "targetNameSource": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "ConfigBatch": {
        "required": [
          "items"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConfigItem"
            }
          }
        },
        "additionalProperties": false
      },
      "ConfigItem": {
        "required": [
          "key",
          "value"
        ],
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "value": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "EmptyResponse": {
        "type": "object",
        "additionalProperties": false
      },
...
      "MailAuditEntry": {
        "required": [
          "action",
          "createdAt",
          "detail",
          "id",
          "targetId",
          "targetType"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "action": {
            "type": "string"
          },
          "targetType": {
            "type": "string",
            "nullable": true
          },
          "targetId": {
            "type": "string",
            "nullable": true
          },
          "detail": {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "MailQueueCounts": {
        "required": [
          "cancelled",
          "failed",
          "pending",
          "sending",
          "sent"
        ],
        "type": "object",
        "properties": {
          "pending": {
            "type": "integer",
            "format": "int64"
          },
          "sending": {
            "type": "integer",
            "format": "int64"
          },
          "sent": {
            "type": "integer",
            "format": "int64"
          },
          "failed": {
            "type": "integer",
            "format": "int64"
          },
          "cancelled": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
      "MailStatusResponse": {
        "required": [
          "configured",
          "from",
          "host",
          "latestFailedAt",
          "latestSentAt",
          "missingEmailAccounts",
          "missingEmailCount",
          "notificationPolicy",
          "notificationsEnabled",
          "port",
          "queue",
          "recent"
        ],
        "type": "object",
        "properties": {
          "configured": {
            "type": "boolean"
          },
          "host": {
            "type": "string",
            "nullable": true
          },
          "port": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "from": {
            "type": "string",
            "nullable": true
          },
          "notificationsEnabled": {
            "type": "boolean"
          },
          "notificationPolicy": {
            "$ref": "#/components/schemas/NotificationPolicyResponse"
          },
          "queue": {
            "$ref": "#/components/schemas/MailQueueCounts"
          },
          "latestSentAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "latestFailedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "missingEmailCount": {
            "type": "integer",
            "format": "int64"
          },
          "missingEmailAccounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MissingEmailAccount"
            }
          },
          "recent": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MailAuditEntry"
            }
          }
        },
        "additionalProperties": false
      },
...
      "MissingEmailAccount": {
        "required": [
          "employeeNo",
          "realName",
          "status",
          "userId",
          "userType"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "employeeNo": {
            "type": "string"
          },
          "realName": {
            "type": "string"
          },
          "userType": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "NotificationEventsResponse": {
        "required": [
          "fileUploaded",
          "messageCreated",
          "projectConfirmed",
          "projectRejected",
          "projectSubmitted",
          "projectWithdrawn"
        ],
        "type": "object",
        "properties": {
          "messageCreated": {
            "type": "boolean"
          },
          "fileUploaded": {
            "type": "boolean"
          },
          "projectSubmitted": {
            "type": "boolean"
          },
          "projectConfirmed": {
            "type": "boolean"
          },
          "projectRejected": {
            "type": "boolean"
          },
          "projectWithdrawn": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "NotificationPolicyResponse": {
        "required": [
          "events",
          "globalEnabled",
          "internalEnabled",
          "supplierEnabled"
        ],
        "type": "object",
        "properties": {
          "globalEnabled": {
            "type": "boolean"
          },
          "internalEnabled": {
            "type": "boolean"
          },
          "supplierEnabled": {
            "type": "boolean"
          },
          "events": {
            "$ref": "#/components/schemas/NotificationEventsResponse"
          }
        },
        "additionalProperties": false
      },
      "PageResponseOfAuditLogResponse": {
        "required": [
          "list",
          "page",
          "pageSize",
          "total"
        ],
        "type": "object",
        "properties": {
          "list": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditLogResponse"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "page": {
            "type": "integer",
            "format": "int64"
          },
          "pageSize": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
...
      "SmtpSettingsUpdate": {
        "required": [
          "from",
          "host",
          "password",
          "port",
          "security",
          "username"
        ],
        "type": "object",
        "properties": {
          "host": {
            "type": "string"
          },
          "port": {
            "type": "integer",
            "format": "int32"
          },
          "username": {
            "type": "string"
          },
          "from": {
            "type": "string"
          },
          "security": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SmtpSettingsView": {
        "required": [
          "configured",
          "from",
          "hasPassword",
          "host",
          "passwordNeedsUpdate",
          "port",
          "security",
          "username"
        ],
        "type": "object",
        "properties": {
          "host": {
            "type": "string"
          },
          "port": {
            "type": "integer",
            "format": "int32"
          },
          "username": {
            "type": "string"
          },
          "from": {
            "type": "string"
          },
          "security": {
            "type": "string"
          },
          "hasPassword": {
            "type": "boolean"
          },
          "configured": {
            "type": "boolean"
          },
          "passwordNeedsUpdate": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "SystemConfigResponse": {
        "required": [
          "description",
          "key",
          "updatedAt",
          "value"
        ],
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "value": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}